requirements_minimal.txt
# Moon Dev AI Agents - Minimal Requirements
# Core packages for research_agent, rbi_agent, and trading_agent

# AI APIs
anthropic>=0.40.0
openai>=1.0.0
google-generativeai
groq

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pandas_ta

# HTTP & Async
requests
aiohttp
httpx
websockets

# Environment & Config
python-dotenv

# Terminal output
termcolor

# Backtesting
backtesting>=0.3.3

# Blockchain / Trading
ccxt>=4.0.0
solana
solders
web3
eth-account

# Utilities
beautifulsoup4
selenium
pydantic

# Technical analysis
ta
